+++ ceammc/CHANGELOG.md
@@ -73,6 +73,7 @@
   - brown, pink, white, blue and velvet methods added to noise.colored~
   - div added to live.capture~ to subdivide loop length while running
   - proto.midi.cc - new portament method added
+  - hexbeat and skip added to seq.bangs
 - new object properties:
   - @punct added to rhvoice~, to specify pronounced characters
   - @bypass added to dyn.comp~ and dyn.comp2~